move storeManager maps and fxns out of main
